refactor: audit Tier 4 structural — api/files + window-tabs splits, fuzzy lift, refresh ownership (#212)
- A5: api/files.ts (1438 → 742 lines) split into api/{common,search,thumbnails,
  archive,config,git,warm-pool}.ts behind a re-export façade; the mock-aware
  invoke seam lives in common.ts; 11 raw-invoke escapees across 5 files now go
  through typed wrappers.
- A4: window-tabs.svelte.ts (1036 → 885 lines): persistence/migration helpers
  → window-tabs-persistence.ts; closed-tab stack → closed-tabs.ts; git-mode
  tab display → domain/tab-title.ts gitTabDisplay (unit-tested). All moved
  symbols re-exported so importers are unchanged.
- A6: ranking math unified in domain/fuzzy-score.ts — filenameMatchScore
  (QuickOpen) and scoreCommand/commandFrecencyPoints (CommandPalette) are now
  pure, shared, and unit-tested; both components consume them.
- A9: refresh-policy ownership declared in one place (refresh-manager.ts
  header + cross-cutting.md table): refresh-manager owns WHEN, pane-watch owns
  WHETHER (mutation cooldown), pane-refresh owns HOW (fingerprint, no flash).
- docs: frontend.md api-layer section updated to the new module layout.

Verified: bun run check 0 errors; 862 unit tests; clippy/fmt clean; Playwright
340 passed (1 perf-budget flake under load, passes in isolation).
